Mod: replace all tab into 2 space
For coffee script which takes advantage of indent to compile the code,
using tab is unbelievable and intolerant, especially for collaboration.
You never know how others set their tab width or weather set expandtab.
It annoyed me so much when I modified the code and debugged for a long
time finding that the code was build without my purpose.